fix pr problems

+++ oneflow/core/common/util.h
@@ -147,7 +147,7 @@ void CloseStdoutAndStderr();
 //  RoundUpToAlignment(bytes, 8) returns 8, if bytes <= 8
 //  RoundUpToAlignment(bytes, 8) returns 16, if 9 < bytes <= 16,
 //  `align' should always be pow(2, (unsigned int)N)
-inline size_t RoundUpToAlignment(size_t bytes, int align) {
+inline size_t RoundUpToAlignment(size_t bytes, size_t align) {
   return (bytes + align - 1) & ~(align - 1);
 }
